fix(servers): clear C4267 size_t narrowing in string and packet paths.
Cast strlen/container sizes explicitly when writing shorts and sizing dstring buffers under /W3.

+++ source/src/groupserver/GroupServerApp.cpp
@@ -280,8 +280,8 @@ void GroupServerApp::SendToAllClients(const WPacket& wpk) {
 			break;
 		}
 
-		g_gpsvr->SendToClient(players.data(), size, wpk);
-		n += size;
+		g_gpsvr->SendToClient(players.data(), static_cast<short>(size), wpk);
+		n += static_cast<int>(size);
 	}
 }
